Witness type coverage: fix Field/Uint truncation, reject Bytes<N>
The transcript codegen was emitting
  `Fr::from(witnesses.<field>.value() as u64)`
for every witness access. The `as u64` cast silently truncated
`Field` and `Uint<N>` values whose `.value()` returns `u128`, losing
the top 64 bits. `Bytes<N>` doesn't have a `.value()` method at all,
so any contract with a `Bytes<N>` witness failed to compile with a
confusing macro-expansion error pointing at the contract module.

Two fixes:

1. `crates/midnight-codegen/src/transcript_codegen.rs`: drop the
   `as u64` cast. `Fr::from(...)` resolves via the `From<bool>`,
   `From<u64>`, `From<u128>` impls in midnight-transient-crypto's
   `curve.rs:268-291`, so the same code path now correctly serializes
   Boolean, Field, and Uint<N> without truncation.

2. `crates/midnight-ir/src/parse.rs::parse_witnesses_struct`: detect
   `Bytes<N>` witness fields and return a clear `MIDNIGHT-001`
   InvalidType error pointing the user at the supported set
   (Boolean/Field/Uint<N>). Multi-Fr witness emission is a real
   refactor (transcript codegen would need to use
   `FieldRepr::field_repr` instead of a single `push`; ZKIR emitter
   would need to emit one PrivateInput per Fr the witness produces).
   The parse-time rejection prevents silently-wrong behavior in the
   meantime.

`crates/midnight/tests/witness_types_test.rs` covers:
- A multi-typed witness struct compiles.
- Each witness type produces a 1-Fr `private_transcript` entry.
- `Field::from(u128::MAX)` survives the round trip — regression check
  for the truncation bug.
- `Bytes<N>` witness produces the documented parse-time error.

`memories/witness-type-support.md` captures the current support matrix
and what multi-Fr witness emission would require. Indexed in
`memories/INDEX.md`.

+++ crates/midnight-codegen/src/transcript_codegen.rs
@@ -206,8 +206,18 @@ fn generate_op_stmt(expr: &ExprIR, field_names: &[String]) -> TokenStream {
         ExprIR::WitnessAccess { field, .. } => {
             let field_ident = format_ident!("{}", field.to_string());
             // Read the witness value and add to private transcript as Fr.
+            //
+            // `Boolean::value()` returns `bool`, `Field::value()` and
+            // `Uint::<N>::value()` both return `u128`. `Fr: From<bool>`,
+            // `From<u64>`, `From<u128>` — so `Fr::from(value())` works for
+            // all three without the previous `as u64` cast, which silently
+            // truncated `Field` and large `Uint<N>` values.
+            //
+            // Multi-Fr witnesses (e.g. `Bytes<N>`) are rejected at parse
+            // time in `crates/midnight-ir/src/parse.rs::parse_witnesses_struct`
+            // until we add proper multi-Fr witness emission.
             quote! {
-                private_transcript.push(Fr::from(witnesses.#field_ident.value() as u64));
+                private_transcript.push(Fr::from(witnesses.#field_ident.value()));
             }
         }
